Separate maven profile extras into three different profiles
With the profile extras it was not possible to build the shaded-jar artefacts without building
the source and the javadoc artefact, this means it was not able to create the standalone jars
on your local system if there where e.g. errors in your javadoc, or if you did not have `pgp(.exe)`
installed to sign the source artefacts.

Now all three profiles are still selected as release profiles, but the standalone jars can
be built independently from the other profiles.